move to 32 bit

+++ Implementierung/benchmark.h
@@ -1,1 +1,3 @@
-void benchmark(long degree, long repetitions);
+#include <stdint.h>
+
+void benchmark(uint32_t degree, uint32_t repetitions);
